Change IPlatform Index property type to ushort

+++ Assets/Scripts/Game/Behaviours/Managers/TwoPlayerLevel.cs
@@ -25,7 +25,7 @@ namespace Game.Behaviours.Managers {
         
         private float highestPlatform;
         
-        private int platformIndex = 0;
+        private ushort platformIndex = 0;
         
         private void OnClientConnected(ulong clientId) {
             InitializeClient();
